created schema graph class
