+++ schematic_db/manifest_store/manifest_store.py
@@ -10,6 +10,8 @@ from schematic_db.api_utils.api_utils import (
     ManifestSynapseConfig,
 )
 
+from schematic_db.schema_graph import SchemaGraph
+
 
 class ManifestMissingPrimaryKeyError(Exception):
     """Raised when a manifest is missing its primary key"""
@@ -93,19 +95,8 @@ class ManifestStore:  # pylint: disable=too-many-instance-attributes
         self.synapse_project_id = config.synapse_project_id
         self.synapse_asset_view_id = config.synapse_asset_view_id
         self.synapse_input_token = config.synapse_input_token
-        self.schema_graph = self.create_schema_graph()
+        self.schema_graph = SchemaGraph(config.schema_url)
         self.update_manifest_configs()
-
-    def create_schema_graph(self) -> networkx.DiGraph:
-        """Retrieve the edges from schematic API and store in networkx.DiGraph()
-
-        Returns:
-            networkx.DiGraph: The edges of the graph
-        """
-        subgraph = get_graph_by_edge_type(self.schema_url, "requiresComponent")
-        schema_graph = networkx.DiGraph()
-        schema_graph.add_edges_from(subgraph)
-        return schema_graph
 
     def create_sorted_object_name_list(self) -> list[str]:
         """
@@ -116,7 +107,7 @@ class ManifestStore:  # pylint: disable=too-many-instance-attributes
         Returns:
             list[str]: A list of objects names
         """
-        return list(reversed(list(networkx.topological_sort(self.schema_graph))))
+        return self.schema_graph.create_sorted_object_name_list()
 
     def update_manifest_configs(self) -> None:
         """Updates the current objects manifest_configs."""
