+++ schematic_db/schema_graph/__init__.py
@@ -1,0 +1,4 @@
+"""
+The ManifestStore class interacts with the Schematic API download manifests.
+"""
+from .schema_graph import *
